### 2025-12-12 #9614 @editorialbot
[PRE REVIEW]: GEFF: Graph Exchange File Format
**Submitting author:** @cmalinmayor (<a href="http://orcid.org/0000-0002-9627-6030">Caroline Malin-Mayor</a>)
**Repository:** https://github.com/live-image-tracking-tools/geff
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.1.3.1.1
