### 2025-12-16 2358ffd
Add conda-forge badges (#385)
# Proposed Change
Closes #363 

Thank you @tlambert03 for helping us get geff and geff-spec up on
conda-forge! 🙏 🎉

# Types of Changes
What types of changes does your code introduce? Delete those that do not
apply.
- Maintenance (e.g. dependencies, CI, releases, etc.)
