+++ joss/paper.bib
@@ -1,4 +1,5 @@
 @article{mavska2014benchmark,
+  doi = {10.1093/bioinformatics/btu080},
   title={A benchmark for comparison of cell tracking algorithms},
   author={Ma{\v{s}}ka, Martin and Ulman, Vladim{\'\i}r and Svoboda, David and Matula, Pavel and Matula, Petr and Ederra, Cristina and Urbiola, Ainhoa and Espa{\~n}a, Tom{\'a}s and Venkatesan, Subramanian and Balak, Deepak MW and others},
   journal={Bioinformatics},
@@ -9,6 +10,7 @@
   publisher={Oxford University Press}
 },
 @article{tinevez2017trackmate,
+  doi = {10.1016/j.ymeth.2016.09.016}, 
   title={TrackMate: An open and extensible platform for single-particle tracking},
   author={Tinevez, Jean-Yves and Perry, Nick and Schindelin, Johannes and Hoopes, Genevieve M and Reynolds, Gregory D and Laplantine, Emmanuel and Bednarek, Sebastian Y and Shorte, Spencer L and Eliceiri, Kevin W},
   journal={Methods},
@@ -26,6 +28,7 @@
   note = {Accessed 2025-11-05}
 },
 @inproceedings{networkx_2008,
+  doi = {10.25080/tcwv9851},
 	address = {Pasadena, CA USA},
 	title = {Exploring {Network} {Structure}, {Dynamics}, and {Function} using {NetworkX}},
 	booktitle = {Proceedings of the 7th {Python} in {Science} {Conference}},
@@ -47,6 +50,7 @@
   journal = {Journal of Open Source Software}
 }
 @article{bragantini2025ultrack,
+  doi = {10.1101/2024.09.02.610652},
   title={Ultrack: pushing the limits of cell tracking across biological scales},
   author={Bragantini, Jord{\~a}o and Theodoro, Ilan and Zhao, Xiang and Huijben, Teun APM and Hirata-Miyasaki, Eduardo and VijayKumar, Shruthi and Balasubramanian, Akilandeswari and Lao, Tiger and Agrawal, Richa and Xiao, Sheng and others},
   journal={Nature Methods},
