### 2025-12-16 63b7317
Add weekly scheduled runs of tests
