### 2025-12-16 8ec5945
Add countries to affiliations
