### 2025-12-16 #9614 @cmalinmayor
Added countries to all university affiliations
