### 2025-12-16 #9614 @danielskatz
> [@editorialbot](https://github.com/editorialbot) generate pdf Added countries to all university affiliations

We don't need cities, just countries, though it's ok to have them too. And we need countries for all affiliations, not just the university ones.
