### 2025-12-17 9a983a5
Add specific city for Yohsuke
