### 2025-12-18 e40a3fb
Change Tobi to Independent Consultant affiliation
