### 2025-12-18 #9614 @cmalinmayor
[Raphtory: The temporal graph engine for Rust and Python](https://github.com/openjournals/joss-reviews/issues/5940)
I think this Raphtory paper found by the bot is quite relevant, as an in-memory graph library that could persist to GEFF
