### 2026-01-15 8240f4e
Add assert to satisfy mypy
