### 2026-01-28 1cb5907
Draft Software Design section
