+++ joss/paper.md
@@ -151,7 +151,7 @@ GEFF’s object specification supports a wide range of shapes, enabling its appl
 The library integrates multiple representation formats, from binary masks (2D/3D), commonly used in cell biology, developmental biology, and natural image tracking—to geometric primitives (points, circles, ellipses, spheres, and ellipsoids), which are essential in super-resolution microscopy, virology, and developmental studies. 
 It also includes polygons and meshes for detailed structural analysis in cell biology, microbiology, and complex shape modeling, as well as pose-based representations for markerless tracking of anatomical keypoints in multi-animal and behavioral research.
 
-Each object track or lineage is represented by a simple directed graph, where each edge is a link from one biological object detected in a frame, to its detection in the earliest next frame. 
+Each object, track, or lineage, is represented by a simple directed graph, where each edge is a link from one biological object detected in a frame, to its detection in the earliest next frame. 
 GEFF can therefore be used to harness tracking data with gaps (an edge extends over more that two adjacent time-points, because of a missing detection or object exit and reentry), object divisions and objects fusions. 
 Multiple GEFFs can refer to the same set of segmentations or detections which makes it possible to easily store multiple tracking solutions or hypotheses without duplicating the underlying data, which is a limitation of the commonly-used CTC format. 
 
@@ -159,11 +159,11 @@ By integrating these diverse object and link representations, GEFF facilitates s
 This versatility makes the library a valuable resource for researchers working across disciplines, from high-throughput cell biology to fine-grained anatomical studies in developmental and computational biology.
 
 ## Software Design
-The GEFF specification emphasizes simplicity over optimization. It would be infeasible to make a single graph format, or even tracking format, that was optimized for all stages of the tracking pipeline from algorithm development to visualization. Implementing a simple exchange format allows a variety of tools with different use cases to all exchange information, while retaining their optimized internal formats. Additionally, focusing on an exchange format allows cross-language compatibility. In addition to a hand-written specification, the GEFF specification package provides Pydantic classes and a JSON schema for creating and validating the metadata properties.
+The GEFF specification emphasizes simplicity over optimization. It would be infeasible to make a single graph format, or even tracking format, that was optimized for all use cases. Implementing a simple exchange format allows a variety of tools with different goals to exchange information, while retaining their optimized internal formats. Additionally, focusing on an exchange format allows cross-language compatibility. In addition to a hand-written specification, the GEFF specification package provides Pydantic classes and a JSON schema for creating and validating the metadata properties.
 
-The Python reference implementation has a highly modular design. The `core_io` module is the heart of the implementation, with core read and write implementations that can be used by any graph library, converting from an `InMemoryGeff` data structure to an on-disk `zarr` and back. This crucial abstraction improves maintainability as well as extensibility; new graph library implementations only need to convert to and from the `InMemoryGeff` data structure, and any improvements to performance or bugfixes in the `core_io` module are automatically propagated to the individual graph library implementations. GEFF supports Zarr specification v2 and v3, and has minimal dependencies, making it a lightweight dependency for other libraries. 
+The Python reference implementation has a highly modular design. The `core_io` module is the heart of the implementation, with core read and write implementations that can be used by any graph library, converting from an `InMemoryGeff` data structure to an on-disk `zarr` and back. This crucial abstraction improves maintainability as well as extensibility; new graph library implementations only need to convert to and from the `InMemoryGeff` data structure, and any improvements to performance or bugfixes in the `core_io` module are automatically propagated to the individual graph library implementations. GEFF supportsZarr specification v2 and v3, and has minimal dependencies, making it a lightweight dependency for other libraries. 
 
[…]
