### 2026-01-29 831872b
Revert "Draft Software Design section"
This reverts commit ae5a3812f9a11e6aa767599698f572cc67f2ba13.
