+++ joss/paper.md
@@ -161,7 +161,7 @@ This versatility makes the library a valuable resource for researchers working a
 ## Software Design
 The GEFF specification emphasizes simplicity over optimization. It would be infeasible to make a single graph format, or even tracking format, that was optimized for all use cases. Implementing a simple exchange format allows a variety of tools with different goals to exchange information, while retaining their optimized internal formats. Additionally, focusing on an exchange format allows cross-language compatibility. In addition to a hand-written specification, the GEFF specification package provides Pydantic classes and a JSON schema for creating and validating the metadata properties.
 
-The Python reference implementation has a highly modular design. The `core_io` module is the heart of the implementation, with core read and write implementations that can be used by any graph library, converting from an `InMemoryGeff` data structure to an on-disk `zarr` and back. This crucial abstraction improves maintainability as well as extensibility; new graph library implementations only need to convert to and from the `InMemoryGeff` data structure, and any improvements to performance or bugfixes in the `core_io` module are automatically propagated to the individual graph library implementations. GEFF supportsZarr specification v2 and v3, and has minimal dependencies, making it a lightweight dependency for other libraries. 
+The Python reference implementation has a highly modular design. The `core_io` module is the heart of the implementation, with core read and write implementations that can be used by any graph library, converting from an `InMemoryGeff` data structure to an on-disk `zarr` and back. This crucial abstraction improves maintainability as well as extensibility; new graph library implementations only need to convert to and from the `InMemoryGeff` data structure, and any improvements to performance or bugfixes in the `core_io` module are automatically propagated to the individual graph library implementations. GEFF supports Zarr specification v2 and v3, and has minimal dependencies, making it a lightweight dependency for other libraries. 
 
 The `_graph_libs` currently implements three graph backends. `networkx` and `rustworkx` are two of the most common Python graph libraries, making GEFF adoption simple for most Python programmers. `spatial_graph` is a newer graph library with improved efficiency when searching for all graph elements in a spatial region. While the Python implementation contains significant internal typing and abstraction logic that reduces code duplication and enhances maintainability, the external API is quite simple; most users will only need to use the public `read` and `write` functions.
 
