### 2026-01-30 cf541a8
Mention prior efforts and how GEFF is different
Co-authored-by: Mark Kittisopikul <mkitti@users.noreply.github.com>
