+++ joss/paper.md
@@ -142,6 +142,7 @@ The developers have already fielded inquiries through GitHub and email from rese
 
 There are many formats used to store and exchange tracking solutions. A commonly used one is the Cell Tracking Challenge (CTC) [@mavska2014benchmark] format, which combines TIFF files with segmentation masks and a CSV file to provide division edges. However, some tracking applications such as particle tracking do not operate on segmentations, but instead utilize point detections, making this format not applicable. As such, individual tracking tools often define their own format for saving tracking results; for example, TrackMate [@tinevez2017trackmate] has a specific XML file format, Mastodon [@tinevez2025mastodon] saves and loads from a binary file, the [Motile Tracker](https://github.com/funkelab/motile_tracker) exports and loads to CSV files with specific node ID, parent ID, and location columns, and Ultrack [@bragantini2025ultrack] has a custom SQL database. In these existing file formats, there is limited support for storing additional properties on either nodes or edges. 
 Additionally, none of these tools shared a common file format, which prevented interactions between them and strongly limited the scope and ambition of track analysis pipelines.
+Additionally, none of these tools shared a common file format, which prevented interactions between them and strongly limited the scope and ambition of track analysis pipelines. In contrast to prior efforts to create a common file format [@gonzalez2020cmso], we worked with tool authors to integrate support for GEFF directly into the tools themselves rather than only creating a third-party library.
 Each of them can now export to and import from GEFF in addition to their custom formats, enabling interoperability with minimal code change in each library.
 
 # Implementation
