+++ joss/paper.bib
@@ -58,6 +58,17 @@
   year={2025},
   publisher={Nature Publishing Group US New York}
 }
+@article{gonzalez2020cmso,
+  doi = {10.1093/gigascience/giaa041},
+  title={Community standards for open cell migration data},
+  author={Gonzalez-Beltran, Alejandra N and Masuzzo, Paola and Ampe, Christophe and Bakker, Gert-Jan and B{\'e}sson, S{\'e}bastien and Eibl, Robert H and Friedl, Peter and Gunzer, Matthias and Kittisopikul, Mark and Le D{\'e}v{\'e}dec, Sylvia E and Leo, Simone and Moore, Josh and Paran, Yael and Prilusky, Jaime and Rocca-Serra, Philippe and Roudot, Philippe and Schuster, Marc and Sergeant, Gwendolien and Str{\"o}mblad, Staffan and Swedlow, Jason R and van Erp, Merijn and Van Troys, Marleen and Zaritsky, Assaf and Sansone, Susanna-Assunta and Martens, Lennart},
+  journal={GigaScience},
+  volume={9},
+  number={5},
+  pages={giaa041},
+  year={2020},
+  publisher={Oxford University Press}
+},
 @software{tinevez2025mastodon,
   title={Mastodon – a large-scale tracking and track-editing framework for large, multi-view images},
   author={Pietzsch, Tobias and Tinevez, Jean-Yves and Sugawara, Ko and Arzt, Matthias and Ulman, Vladimir and Hahmann, Stefan},
