### 2026-01-30 57c7a4b
Add acknowledgments and competing interests for KS
