### 2026-01-30 73f4461
Add BD5 citation
