+++ joss/paper.bib
@@ -76,3 +76,17 @@
   url = {https://mastodon.readthedocs.io/en/latest/},
   note = {Accessed 2025-11-12}
 },
+@article{kyoda2020bd5,
+    doi = {10.1371/journal.pone.0237468},
+    author = {Kyoda, Koji AND Ho, Kenneth H. L. AND Tohsato, Yukako AND Itoga, Hiroya AND Onami, Shuichi},
+    journal = {PLOS ONE},
+    publisher = {Public Library of Science},
+    title = {BD5: An open HDF5-based data format to represent quantitative biological dynamics data},
+    year = {2020},
+    month = {08},
+    volume = {15},
+    url = {https://doi.org/10.1371/journal.pone.0237468},
+    pages = {1-11},
+    abstract = {BD5 is a new binary data format based on HDF5 (hierarchical data format version 5). It can be used for representing quantitative biological dynamics data obtained from bioimage informatics techniques and mechanobiological simulations. Biological Dynamics Markup Language (BDML) is an XML (Extensible Markup Language)-based open format that is also used to represent such data; however, it becomes difficult to access quantitative data in BDML files when the file size is large because parsing XML-based files requires large computational resources to first read the whole file sequentially into computer memory. BD5 enables fast random (i.e., direct) access to quantitative data on disk without parsing the entire file. Therefore, it allows practical reuse of data for understanding biological mechanisms underlying the dynamics.},
+    number = {8},
+},
