### 2026-02-03 #9614 @danielskatz
👋 @gkthiruvathukal - as the editor, can you see if you think this meets the current JOSS scope, and if so, could you get it moving?  I'll remove the paused label.
