### 2026-02-17 #9614 @gkthiruvathukal
👋 [antoineruzette](https://github.com/antoineruzette) and/or
[bencardoen](https://github.com/bencardoen) - Would you be willing to review this submission to JOSS? I need a second reviewer to proceed with review. We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
