+++ .github/workflows/ci.yaml
@@ -153,7 +153,7 @@ jobs:
       contents: write # this permission is required for generate_release_notes
     steps:
       - name: Download built artifact to dist/
-        uses: actions/download-artifact@v7
+        uses: actions/download-artifact@v8
         with:
           name: Packagesgeff
           path: dist
@@ -194,7 +194,7 @@ jobs:
       contents: write # this permission is required for generate_release_notes
     steps:
       - name: Download built artifact to dist/
-        uses: actions/download-artifact@v7
+        uses: actions/download-artifact@v8
         with:
           name: Packagesgeff-spec
           path: dist
