### 2026-03-04 #10143 @editorialbot
[REVIEW]: GEFF: Graph Exchange File Format
**Submitting author:** @cmalinmayor (<a href="http://orcid.org/0000-0002-9627-6030">Caroline Malin-Mayor</a>)
**Repository:** https://github.com/live-image-tracking-tools/geff
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.1.5.1.1
**Archive:** 10.25378/janelia.31943145
