### 2026-03-16 c499fff
Include a simple quickstart in the readme
