### 2026-03-18 41624f2
CI fix: Clear entire temp dir and not just dist directory (#401)
Per Claude, our CI step:

> 1. Uses /tmp/baipp as its working directory for the virtual
environment
> 2. The action is being called twice (once for geff-spec at line 135,
once for geff at line 141)
> 3. Between the two calls, you're clearing the dist directory (line
140) but not the virtual environment
> 
> The error message shows that when the action tries to create a venv at
/tmp/baipp for the second package build, it fails because the venv from
the first build still exists. The fix is to clear the entire /tmp/baipp
directory b
[…]
