+++ .github/workflows/ci.yaml
@@ -136,8 +136,8 @@ jobs:
         with:
           path: packages/geff-spec
           upload-name-suffix: geff-spec
-      - name: Clear dist output directory
-        run: rm -r /tmp/baipp/dist
+      - name: Clear build directory
+        run: rm -rf /tmp/baipp
       - uses: hynek/build-and-inspect-python-package@v2
         with:
           path: packages/geff
