### 2026-03-18 #10143 @msschwartz21
@Pablo1990 Thank you for your feedback! Both of the issues you opened have been addressed and those changes have been merged back into the `joss` branch.
