### 2026-03-30 93a96d0
check for booleans before checking for integers when determining the default value
