+++ .github/workflows/ci.yaml
@@ -81,7 +81,7 @@ jobs:
           uv run coverage xml
 
       - name: Coverage
-        uses: codecov/codecov-action@v5
+        uses: codecov/codecov-action@v6
         with:
           token: ${{ secrets.CODECOV_TOKEN }}
 
