### 2026-03-31 98f00e7
Update docstrings and tests
