### 2026-03-31 a4dff33
Refactor dtype determination for reuse in arrays or dicts
