### 2026-03-31 fb1cfba
Update default value determination to handle numpy types
