### 2026-03-31 e0401f3
Add dataframe to geff
