+++ packages/geff/src/geff/convert/__init__.py
@@ -7,6 +7,8 @@ if TYPE_CHECKING:
 
 __all__ = [
     "ctc_tiffs_to_zarr",
+    "dataframes_to_geff",
+    "dataframes_to_memory_geff",
     "from_ctc_to_geff",
     "from_trackmate_xml_to_geff",
     "geff_to_csv",
@@ -27,6 +29,14 @@ def __getattr__(name: str) -> Any:
         from geff.convert._dataframe import geff_to_csv
 
         return geff_to_csv
+    if name == "dataframes_to_geff":
+        from geff.convert._dataframe import dataframes_to_geff
+
+        return dataframes_to_geff
+    if name == "dataframes_to_memory_geff":
+        from geff.convert._dataframe import dataframes_to_memory_geff
+
+        return dataframes_to_memory_geff
     if name == "geff_to_dataframes":
         from geff.convert._dataframe import geff_to_dataframes
 
