+++ packages/geff/src/geff/convert/__init__.py
@@ -6,6 +6,7 @@ if TYPE_CHECKING:
     from ._ctc import ctc_tiffs_to_zarr, from_ctc_to_geff
 
 __all__ = [
+    "csv_to_geff",
     "ctc_tiffs_to_zarr",
     "dataframes_to_geff",
     "dataframes_to_memory_geff",
@@ -25,6 +26,10 @@ def __getattr__(name: str) -> Any:
         from geff.convert._ctc import from_ctc_to_geff
 
         return from_ctc_to_geff
+    if name == "csv_to_geff":
+        from geff.convert._dataframe import csv_to_geff
+
+        return csv_to_geff
     if name == "geff_to_csv":
         from geff.convert._dataframe import geff_to_csv
 
