### 2026-03-31 5b3ccd1
Try to get new functions to show up in docs
