+++ packages/geff/src/geff/core_io/__init__.py
@@ -1,9 +1,10 @@
 from ._base_read import read_to_memory
 from ._base_write import write_arrays, write_dicts
-from ._utils import check_for_geff, construct_var_len_props, delete_geff
+from ._utils import check_for_geff, construct_props, construct_var_len_props, delete_geff
 
 __all__ = [
     "check_for_geff",
+    "construct_props",
     "construct_var_len_props",
     "delete_geff",
     "read_to_memory",
