### 2026-04-02 5497a46
Remove assumption about index column in csvs
