### 2026-04-06 48e9018
ci(pre-commit.ci): autoupdate
updates:
- [github.com/adhtruong/mirrors-typos: v1.42.3 → v1.45.0](https://github.com/adhtruong/mirrors-typos/compare/v1.42.3...v1.45.0)
- [github.com/astral-sh/ruff-pre-commit: v0.14.14 → v0.15.9](https://github.com/astral-sh/ruff-pre-commit/compare/v0.14.14...v0.15.9)
- [github.com/abravalheri/validate-pyproject: v0.24.1 → v0.25](https://github.com/abravalheri/validate-pyproject/compare/v0.24.1...v0.25)
- [github.com/pre-commit/mirrors-mypy: v1.19.1 → v1.20.0](https://github.com/pre-commit/mirrors-mypy/compare/v1.19.1...v1.20.0)
