### 2026-04-06 #10143 @cmalinmayor
https://doi.org/10.25378/janelia.31943145
geff-spec v1.1.1
geff python library v1.1.5
full github repository v1.1.5.1.1
