### 2026-04-07 435d5bf
Update docstrings for csv to geff conversion
