### 2026-04-07 d5f20ae
Detect and use the smallest int dtype
