+++ packages/geff/tests/test_convert/test_dataframe.py
@@ -5,6 +5,7 @@ try:
     import pandas as pd
 
     from geff.convert._dataframe import (
+        _infer_int_dtype,
         csv_to_geff,
         dataframes_to_geff,
         dataframes_to_memory_geff,
@@ -22,6 +23,40 @@ import pytest
 from geff import _path
 from geff.core_io._base_read import read_to_memory
 from geff.testing.data import create_mock_geff, create_simple_2d_geff, create_simple_3d_geff
+
+
+class Test_infer_int_dtype:
+    def test_small_unsigned(self):
+        values = np.array([0, 1, 255])
+        assert _infer_int_dtype(values) == np.dtype(np.uint8)
+
+    def test_uint16(self):
+        values = np.array([0, 256])
+        assert _infer_int_dtype(values) == np.dtype(np.uint16)
+
+    def test_uint32(self):
+        values = np.array([0, 2**16])
+        assert _infer_int_dtype(values) == np.dtype(np.uint32)
+
+    def test_uint64(self):
+        values = np.array([0, 2**32])
+        assert _infer_int_dtype(values) == np.dtype(np.uint64)
+
+    def test_negative_int8(self):
+        values = np.array([-1, 0, 127])
+        assert _infer_int_dtype(values) == np.dtype(np.int8)
+
+    def test_negative_int16(self):
+        values = np.array([-129, 0])
+        assert _infer_int_dtype(values) == np.dtype(np.int16)
+
+    def test_negative_int32(self):
+        values = np.array([-(2**15) - 1, 0])
+        assert _infer_int_dtype(values) == np.dtype(np.int32)
+
+    def test_negative_int64(self):
+        values = np.array([-(2**31) - 1, 0])
+        assert _infer_int_dtype(values) == np.dtype(np.int64)
 
 
 class Test_geff_to_dataframes:
