### 2026-04-07 8864bbb
Make default_for_value a public utils function
