+++ packages/geff/src/geff/core_io/_base_write.py
@@ -8,9 +8,9 @@ import numpy as np
 from geff import _path
 from geff._typing import PropDictNpArray
 from geff.core_io._utils import (
-    _default_for_value,
     check_for_geff,
     construct_var_len_props,
+    default_for_value,
     delete_geff,
     remove_tilde,
     setup_zarr_group,
@@ -111,8 +111,12 @@ def write_dicts(
 def _determine_default_value(data: Sequence[tuple[Any, dict[str, Any]]], prop_name: str) -> Any:
     """Determine default value to fill in missing values
 
-    Find the first non-missing value and return a type-appropriate default
-    using _default_for_value.
+    Uses the following heuristics:
+    - np.generic (np.bool_, np.int64, np.float32, etc.) -> type(value)(0)
+    - bool, int, float -> type(value)(0) (e.g. False, 0, 0.0)
+    - str -> ""
+    - Otherwise, returns the value itself, which preserves type and shape but
+      may be confusing or inefficient for some types.
 
     If there are no non-missing values, warns and then returns 0.
 
@@ -128,7 +132,7 @@ def _determine_default_value(data: Sequence[tuple[Any, dict[str, Any]]], prop_na
     for _, data_dict in data:
         # find first non-missing value
         if prop_name in data_dict:
-            return _default_for_value(data_dict[prop_name])
+            return default_for_value(data_dict[prop_name])
     warnings.warn(
         f"Property {prop_name} is not present on any graph elements. Using 0 as the default.",
         stacklevel=2,
