### 2026-04-07 a5c4386
Update csv_to_geff to use hand-crafted fixtures
