+++ packages/geff/tests/test_convert/test_dataframe.py
@@ -7,7 +7,6 @@ try:
     from geff.convert._dataframe import (
         _infer_int_dtype,
         csv_to_geff,
-        dataframes_to_geff,
         dataframes_to_memory_geff,
         geff_to_csv,
         geff_to_dataframes,
@@ -241,61 +240,62 @@ class Test_geff_to_csv:
         assert "z" not in df.columns
 
 
+@pytest.fixture()
+def sample_dataframes():
+    """Node and edge DataFrames covering multiple dtypes and missing values."""
+    node_df = pd.DataFrame(
+        {
+            "id": [0, 1, 2, 3, 4, 5],
+            "x": [1.0, 2.0, 3.0, 4.0, 5.0, 6.0],
+            "label": ["a", "b", "c", "d", "e", "f"],
+            "flag": pd.array([True, False, pd.NA, True, False, True], dtype=pd.BooleanDtype()),
+            "score": [10.0, np.nan, 30.0, 40.0, 50.0, 60.0],
+        }
+    )
+    edge_df = pd.DataFrame(
+        {
+            "source": [0, 1, 2, 3],
+            "target": [1, 2, 3, 4],
+            "weight": [0.1, 0.2, 0.3, 0.4],
+        }
+    )
+    return node_df, edge_df
+
+
+@pytest.fixture()
+def expected_memory_geff():
+    """Manually constructed expected InMemoryGeff for the sample DataFrames."""
+    return {
+        "node_ids": np.array([0, 1, 2, 3, 4, 5], dtype=np.uint8),
+        "edge_ids": np.array([[0, 1], [1, 2], [2, 3], [3, 4]], dtype=np.uint8),
+        "node_props": {
+            "x": {
+                "values": np.array([1.0, 2.0, 3.0, 4.0, 5.0, 6.0]),
+                "missing": None,
+            },
+            "label": {
+                "values": np.array(["a", "b", "c", "d", "e", "f"]),
+                "missing": None,
+            },
+            "flag": {
+                "values": np.array([True, False, False, True, False, True]),
+                "missing": np.array([False, False, True, False, False, False]),
+            },
+            "score": {
+                "values": np.array([10.0, 0.0, 30.0, 40.0, 50.0, 60.0]),
+                "missing": np.array([False, True, False, False, False, False]),
+            },
+        },
+        "edge_props": {
+            "weight": {
+                "values": np.array([0.1, 0.2, 0.3, 0.4]),
+                "missing": None,
+            },
+        },
+    }
+
+
 class Test_dataframes_to_geff:
-    def test_round_trip(self):
-        """geff -> dataframes -> geff should preserve data."""
-        store, original = create_simple_3d_geff()
-        node_df, edge_df = geff_to_dataframes(store)
-
-        result = dataframes_to_memory_geff(node_df, edge_df, directed=original["metadata"].directed)
-
-        np.testing.assert_array_equal(result["node_ids"], original["node_ids"])
-        np.testing.assert_array_equal(result["edge_ids"], original["edge_ids"])
-        for name in original["node_props"]:
-            np.testing.assert_array_equal(
-                result["node_props"][name]["values"],
-                original["node_props"][name]["values"],
-            )
-        for name in original["edge_props"]:
-            np.testing.assert_array_equal(
-                result["edge_props"][name]["values"],
-                original["edge_props"][name]["values"],
-            )
-
-    def test_missing_values(self):
-        """NaN values in DataFrame should produce missing masks."""
-        node_df = pd.DataFrame(
-            {
-                "id": [0, 1, 2],
-                "score": [1.0, np.nan, 3.0],
-            }
-        )
[…]
