### 2026-04-07 42ea629
Properly test missing values in all cases
