### 2026-04-07 8dcba8f
Bump pandas lower bound to avoid to_csv bug
