### 2026-04-08 a15cee5
Update pyright settings to point to correct repo locations (#410)
# Proposed Change
Stale configuration in our root pyproject was messing up my
pylance/VSCode imports and autocompletion. Fixed by pointing to the
correct source locations of the two packages within the monorepo.

# Types of Changes
What types of changes does your code introduce? Delete those that do not
apply.
- Maintenance (e.g. dependencies, CI, releases, etc.)

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here 
[…]
