+++ packages/geff/src/geff/convert/__init__.py
@@ -4,6 +4,7 @@ from geff.convert._trackmate_xml import from_trackmate_xml_to_geff
 
 if TYPE_CHECKING:
     from ._ctc import ctc_tiffs_to_zarr, from_ctc_to_geff
+    from ._dataframe import geff_to_csv, geff_to_dataframes
 
 __all__ = [
     "ctc_tiffs_to_zarr",
