### 2026-04-09 f69cf93
Use oindex for sparse zarr reads and fix varlength DATA masking
Replace boolean mask indexing (mask.tolist()) with integer-indexed
oindex reads in _load_prop_to_memory and build(). For sparse
selections (e.g. 5 nodes out of 8674), oindex is ~7x faster because
zarr only reads the chunks containing the requested indices.

Also fix a bug where the varlength DATA array (a flat 1D concatenation
of all variable-length values) was incorrectly indexed with a
per-element boolean mask. DATA must always be loaded in full since
deserialization uses byte offsets from the VALUES array.
