### 2026-04-09 4c5ddf6
Use explicit chunking and zarr v3 sharding when writing arrays
Add _write_zarr_array helper that chunks only along the first
dimension with a power-of-two size targeting ~8 MiB per chunk.
Trailing dimensions are kept whole, fixing zarr auto-chunking that
would unnecessarily split small dimensions (e.g. edge_ids (N, 2)
was chunked as (47468, 1) instead of keeping both columns together).

On zarr v3, sharding is used so each array becomes a single file
with ~8 MiB inner sub-chunks, reducing the file count on disk by
~44% for typical GEFF files.
