### 2026-04-09 c95eab1
Add regression test for varlength properties with node masking
Verifies that _load_prop_to_memory correctly handles varlength
properties when a node mask is applied: DATA must be loaded in full
since it cannot be indexed by a per-node boolean mask.
