### 2026-04-11 75f92c9
Made dataframes to memory function private
