### 2026-04-11 7e968a9
Refactor logic for empty nodes df
Co-authored-by: Morgan Schwartz <msschwartz21@gmail.com>
