### 2026-04-11 3249320
Remove unnecessary if condition
Co-authored-by: Morgan Schwartz <msschwartz21@gmail.com>
