### 2026-04-11 379fa90
Use smallest integer dtype for int props
