### 2026-04-11 8012b23
Upgrade min pandas to 2.2
