### 2026-04-11 71d1975
Ignore pandas pyarrow dependency warning
