### 2026-04-11 e98f2c4
Match newline at start of pyarrow warning
