### 2026-04-11 aef1d12
Go back to pandas 2, ignore the warning
