+++ pyproject.toml
@@ -97,6 +97,7 @@ filterwarnings = [
     "ignore:.*skipping conversion test:ImportWarning:networkx",
     "ignore:In the future `np.bool` will be defined as the corresponding NumPy scalar.",
     "ignore:(?s).*Pyarrow will become a required dependency:DeprecationWarning",
+    "ignore:invalid value encountered in cast:RuntimeWarning",
 ]
 
 
