### 2026-04-13 948efaf
minor changes in paper.md
