### 2026-04-13 #10143 @danielskatz
👋 @cmalinmayor - I've suggested some minor changes in https://github.com/live-image-tracking-tools/geff/pull/413 - please merge this, or let me know what you disagree with.
