### 2026-04-13 c65f19d
minor changes for JOSS publication (#413)
# Proposed Change
Briefly describe the contribution. If it resolves an issue or feature
request, be sure to link to that issue.

# Types of Changes
What types of changes does your code introduce? Delete those that do not
apply.
- Bugfix (non-breaking change which fixes an issue)
- New feature or enhancement
- Documentation update
- Tests
- Maintenance (e.g. dependencies, CI, releases, etc.)

Which topics does your change affect? Delete those that do not apply.
- Specification
- Implementations (`geff._graph_libs`)
- Convert
- Core io
- Metadata
- Testi
[…]
