+++ joss/paper.bib
@@ -30,8 +30,8 @@
 @inproceedings{networkx_2008,
   doi = {10.25080/tcwv9851},
 	address = {Pasadena, CA USA},
-	title = {Exploring {Network} {Structure}, {Dynamics}, and {Function} using {NetworkX}},
-	booktitle = {Proceedings of the 7th {Python} in {Science} {Conference}},
+	title = {Exploring Network Structure, Dynamics, and Function using {NetworkX}},
+	booktitle = {Proceedings of the 7th {P}ython in Science Conference},
 	author = {Hagberg, Aric A. and Schult, Daniel A. and Swart, Pieter J.},
 	editor = {Varoquaux, Gaël and Vaught, Travis and Millman, Jarrod},
 	year = {2008},
@@ -46,7 +46,7 @@
   number = {79}, 
   pages = {3968}, 
   author = {Matthew Treinish and Ivan Carvalho and Georgios Tsilimigkounakis and Nahum Sá}, 
-  title = {rustworkx: A High-Performance Graph Library for Python}, 
+  title = {rustworkx: A High-Performance Graph Library for {P}ython}, 
   journal = {Journal of Open Source Software}
 }
 @article{bragantini2025ultrack,
