### 2026-04-13 #10143 @danielskatz
Congratulations to @cmalinmayor (Caroline Malin-Mayor) and co-authors on your JOSS publication!!

JOSS completely depends on volunteers; we couldn't do this without you
